:fire:style: remove signup info checking display 회원가입 정보 확인 화면 삭제 Related to: #55
